upd: applied re-export pattern
